Reorganized graphics files
