+++ src/components/game/graphics/sections/Stage.js
@@ -1,6 +1,6 @@
-import { STAGE_WIDTH, WINDOW_SIZE } from "../Constants";
-import { drawCell } from "./DrawCell";
-import { drawRect, outlineRect } from "./Shapes";
+import { STAGE_WIDTH, WINDOW_SIZE } from "../../Constants";
+import { drawCell } from "../sprites/DrawCell";
+import { drawRect, outlineRect } from "../utils/Shapes";
 
 const STAGE_CELL_SIZE = WINDOW_SIZE / 32;
