+++ src/components/game/graphics/sprites/DrawPiece.js
