+++ src/components/game/graphics/sprites/UpdateCell.js
